% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_codes.R
\name{add_codes}
\alias{add_codes}
\title{Add More ONS Codes for an area type}
\usage{
add_codes(uk_codes, type, within = NULL)
}
\arguments{
\item{uk_codes}{Either a vector of ONS codes or a dataframe returned by uk_codes()}

\item{type}{The area type to add codes for.}

\item{within}{Limit the codes returned to within a broader area. E.g. Local authorities within the North East}
}
\value{

}
\description{
Add More ONS Codes for an area type
}
