% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_codes.R
\name{find_area_name}
\alias{find_area_name}
\title{Return the Area Name for a given Area Code}
\usage{
find_area_name(area_code)
}
\arguments{
\item{area_code}{An ONS Area Code}
}
\value{
The area name
}
\description{
Return the Area Name for a given Area Code
}
