% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_codes.R
\name{find_area_names}
\alias{find_area_names}
\title{Return the area names for given area codes}
\usage{
find_area_names(area_codes)
}
\arguments{
\item{area_codes}{The ONS area codes to find area names for}
}
\value{
A dataframe of area names and the corresponding area code
}
\description{
Return the area names for given area codes
}
