% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_map.R
\name{get_boundary}
\alias{get_boundary}
\title{Get the boundary data for a single area}
\usage{
get_boundary(area_code)
}
\arguments{
\item{area_code}{An ONS area code}
}
\value{
The boundary data for the area code with class "sfc"
}
\description{
Get the boundary data for a single area
}
