% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ons_query}
\alias{ons_query}
\title{Internal function for making SPARQL queries to gov API}
\usage{
ons_query(query)
}
\arguments{
\item{query}{A SPARQL query}
}
\value{
A dataframe of the results
}
\description{
Internal function for making SPARQL queries to gov API
}
