% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_codes.R
\name{remove_codes}
\alias{remove_codes}
\title{Remove code from selection}
\usage{
remove_codes(uk_codes, area_codes)
}
\arguments{
\item{uk_codes}{Either a vector of ONS codes or a dataframe returned by uk_codes()}

\item{area_codes}{The ONS codes to remove}
}
\value{

}
\description{
Remove code from selection
}
