% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_codes.R
\name{uk_codes}
\alias{uk_codes}
\title{Get ONS area codes for UK area types}
\usage{
uk_codes(type)
}
\arguments{
\item{type}{The Area Collection Code to get area codes for. E.g. "E09" for London Boroughs. All possible collection codes area available at http://statistics.data.gov.uk/atlas}
}
\value{
A character vector of area codes
}
\description{
Get ONS area codes for UK area types
}
\examples{
uk_codes("E09")
}
