% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_codes.R
\name{uk_entity_search}
\alias{uk_entity_search}
\title{Find names and codes of UK area entities}
\usage{
uk_entity_search(term = "", exact = FALSE)
}
\arguments{
\item{term}{Either an area entity code or and an entity name}

\item{exact}{If true then only exact matches to a search term will be returned and will be returned as a single string. If false then any regular expression matches to the search term will be returned.}
}
\value{
If an area code is given then the area name will be returned, if a character string is given then a tibble of entity codes and the entity names that match the string will be returned.
}
\description{
Find names and codes of UK area entities
}
\examples{
uk_entity_search("E09") # returns the string "London Boroughs"

uk_entity_search("London Boroughs", exact = TRUE) # return the string "E09"

uk_entity_search("Super Output Areas") # returns a tibble where entity name matches the search term

}
