% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_map.R
\name{uk_map}
\alias{uk_map}
\title{Get map data for UK Area}
\usage{
uk_map(area_codes)
}
\arguments{
\item{area_codes}{The ONS area codes to be included in the map.}
}
\value{
Boundary data for eaech area code
}
\description{
Get map data for UK Area
}
\examples{

uk_map(uk_codes("E09"))

}
