% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uk_unemployment}
\alias{uk_unemployment}
\title{UK Unemployment Data}
\format{
A data frame with 372 rows and 3 variables:
\describe{
  \item{area_name}{the local authority name}
  \item{area_code}{the ons area code}
  \item{unemployment_rate}{percent of people unemployed}
  ...
}
}
\source{
\url{ONS}
}
\usage{
uk_unemployment
}
\description{
A dataset containing UK unemployment data for local authorities.
}
\keyword{datasets}
