% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_uk.R, R/uk_map.R
\name{uk_map}
\alias{uk_map}
\title{Title}
\usage{
uk_map(area_codes)

uk_map(area_codes)
}
\arguments{
\item{area_codes}{The ONS area codes to be included in the map.}

\item{region}{}
}
\value{



}
\description{
Title

Get map data for UK Area
}
